Split more functions out of the monolithic Verification module
