+++ pate.cabal
@@ -76,8 +76,11 @@ library
                        Pate.Config,
                        Pate.Discovery,
                        Pate.Verification,
+                       Pate.Verification.Domain,
                        Pate.Verification.ExternalCall,
+                       Pate.Verification.Simplify,
                        Pate.Verification.SymbolicExecution,
+                       Pate.Verification.Validity,
                        Pate.Loader,
                        Pate.Loader.ELF,
                        Pate.Event,
